mir_navigation: Remove parameter first_map_only
This parameter must be set to false (the default) when running SLAM
(otherwise the map updates won't be received), and when running a static
map_server it doesn't matter; even then, it should be false to allow
restarting the map_server with a different map. Therefore this commit
removes it altogether and leaves it at the default of "false".