Fix translation error in odom_comb (#12)
Previously, the ekf localization only computed a correct orientation, but the translation still followed the pure odometry data. This led to strange errors where the robot would move sideways (despite only having a diff drive).

This PR changes the ekf configuration to not use any position information from the odometry, but to integrate the velocities, which fixes this problem.