added PathDistPrunedCritic for dwb
which works exactly like the original PathDistCritic, except that it
searches for a local minimum in the distance from the global path to the robots
current position. It then prunes the global_path from the start up to
this point, therefore approximately cutting of a segment of the path
that the robot already followed.