Fix laser scan frame_id with gazebo_plugins 2.9.2
In gazebo_plugins 2.9.2, the behavior of the gazebo_ros_laser plugin was
changed in this commit:

https://github.com/ros-simulation/gazebo_ros_pkgs/commit/34b2c27936cf1b00801b774935afb7a8e17e9c8c

It now no longer prepends the robotNamespace if the parameter tf_prefix
is empty. Since tf_prefix has been deprecated for a long time, we always
leave it empty.

The change in gazebo_plugins means that #65 no longer happens, and we
have to prepend the prefix ourselves (just like with the other plugins).

This reverts commit ba02a68e and applies the following additional diff:

-          <!-- no prefix, plugin always uses a tf_prefix (if none is set it uses robotNamespace) -->
-          <frameName>${link}</frameName>
+          <frameName>${prefix}${link}</frameName>