Increase queue_size for publishers + subscribers to 10
One case where this was a problem was the tf_static topic: Since
multiple messages are being published at once, the subscriber often
missed one. The tf_static topic will be fixed anyway in the next commit,
but let's increase the queue_size anyway to avoid such bugs in the
future.